% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingDatFiles.R
\docType{class}
\name{CtrDat-class}
\alias{CtrDat-class}
\title{An S4 class to represent a .dat file}
\description{
An S4 class to represent a .dat file
}
\section{Slots}{

\describe{
\item{\code{EmployerID}}{The Employer ID for all data in the object}

\item{\code{Validate}}{True if it passes validation, error message if it does not}

\item{\code{Header}}{A one row data frame with header data}

\item{\code{SurveyData}}{A data frame containing the survey data}

\item{\code{FillerRemoved}}{A logical value indicating if filler has been removed}
}}

