% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingDatFiles.R
\name{dats.split}
\alias{dats.split}
\title{Splits all .dat files in folder and moves to destination}
\usage{
dats.split(SourceFolder, DestinationFolder, Validate = TRUE)
}
\arguments{
\item{SourceFolder}{The folder with .dat files needing splitting}

\item{DestinationFolder}{The folder where split .dat files should be written to}

\item{Validate}{Boolean indicating if validation should be run}
}
\description{
Splits all .dat files in the specified folder, by Employer ID. Spilt files are saved into separate .dat
files in the specified destination folder.  It also runs a the validation checks describe in
\code{\link{validate.dat.header}}.
}
\examples{
dats.split("c:/example/source/", "c:/example/destination")
}
