% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingDatFiles.R
\name{read.dat.getEmployers}
\alias{read.dat.getEmployers}
\title{Get list of employers in .dat file}
\usage{
read.dat.getEmployers(Path)
}
\arguments{
\item{Path}{File path to a .dat file}
}
\value{
A list of employer IDs in a .dat file
}
\description{
LOWER LEVEL FUNCTION - Not meant to be called on its own. \cr \cr
Returns a list of employers in a .dat file. This is a helper function around \code{\link{read.dat.headers}}
}
\examples{
read.dat.getEmployers("c:/example/Datfile.dat")
}
