% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingDatFiles.R
\name{read.dat.headers}
\alias{read.dat.headers}
\title{Reads the headers of a .dat file}
\usage{
read.dat.headers(Path, RemoveFiller = TRUE, OnlyFirstHeader = FALSE)
}
\arguments{
\item{Path}{File path to a .dat file}

\item{RemoveFiller}{A boolean indicating if the filler fields should be removed}

\item{OnlyFirstHeader}{A boolean indicating if only the first line of the header should be returned.  Useful for
working with .dat files with only one employer ID.}
}
\value{
A data frame with information formatted using the header schema from a .dat file
}
\description{
LOWER LEVEL FUNCTION - Not meant to be called on its own.\cr \cr
It returns a data frame with .dat data formatted for headers.  All data,
including survey data, will be in the data frame. This may be used to
pull a list of all employer IDs in a .dat file.
}
\examples{
read.dat.header("c:/example/example/datfile.dat")
}
