% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingDatFiles.R
\name{read.dat}
\alias{read.dat}
\title{Reads survey and header data from .dat file with one Employer Id}
\usage{
read.dat(Path, RemoveFiller = TRUE)
}
\arguments{
\item{Path}{File path to a .dat file}

\item{RemoveFiller}{A boolean indicating if the filler fields should be removed}
}
\value{
An S4 object, \code{\link{CtrDat-class}}, with the Employer Id, information from the header of the dat file,
the survey data, a logical value indicating if filler has been removed.
}
\description{
Returns an R object, of type \code{\link{CtrDat-class}}, with survey and header data from a .dat file.
It also preforms validation as described in \code{\link{validate.dat.header}}
}
\examples{
read.dat("c:/example/datfile.dat")
}
