% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingDatFiles.R
\name{read.dat.survey}
\alias{read.dat.survey}
\title{Reads survey data from .dat file}
\usage{
read.dat.survey(Path, RemoveFiller = TRUE)
}
\arguments{
\item{Path}{File path to a .dat file}

\item{RemoveFiller}{A boolean indicating if the filler fields should be removed}
}
\value{
A data frame with information from the header of the .dat file
}
\description{
LOWER LEVEL FUNCTION - Not meant to be called on its own.\cr \cr
Returns a data frame with .dat data formated using the survey schema
}
\examples{
read.dat.survey("c:/example/example/datfile.dat")
}
