% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingDatFiles.R
\name{single.dat.split}
\alias{single.dat.split}
\title{Splits one .dat file by employer ID}
\usage{
single.dat.split(Path, DestinationFolder)
}
\arguments{
\item{Path}{File path to a .dat file}

\item{DestinationFolder}{The folder where split .dat files should be written to}
}
\description{
Splits one .dat file by Employer ID into separate .dat files.  If there is only one employer in
the .dat file, it will be copied to the destination folder.
}
\examples{
single.dat.split("c:/example/datfile.dat", "c:/example/split/")
}
