% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingDatFiles.R
\docType{data}
\name{SurveySchema}
\alias{SurveySchema}
\title{Survey Schema}
\format{A data frame with 67 rows and 7 variables}
\usage{
data(SurveySchema)
}
\description{
This data frame contains the schema for the suvey of fix width file formated .dat files. It contains:
\itemize{
  \item Field.name: The name of the field in the .dat file
  \item field.type: The R data type used to read in the field (set to character to account for leading zeros and spaces)
  \item length: The fixed character length of the field
  \item start: The starting character position of the field
  \item end: The ending character position of the field
  \item description: The field description
  \item HasData: A flag indicating if the field contains data or is filler
}
}
\keyword{datasets}
