% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingDatFiles.R
\name{validate.dat.header}
\alias{validate.dat.header}
\title{Validates header data}
\usage{
validate.dat.header(Header)
}
\arguments{
\item{Header}{a valid header returned from the \code{\link{read.dat.header}} function}
}
\value{
Returns TRUE if it passes validation and an error message if it does not
}
\description{
LOWER LEVEL FUNCTION - Not meant to be called on its own.\cr \cr
Preforms various checks on the header to assure the .dat file meets specifications  The following checks are preformed:\cr
\itemize{
  \item Checks that there is only one valid header in the .dat file
  \item The header flag is equal to 'Y'
  \item Returned surveys are less than or equal to the number distributed
  \item Returned surveys are less than or equal the number of total employees
  \item Distributed surveys are less than or equalt to the number of total employees - exempted employees
}
}
\examples{
validate.dat.header(read.dat.header("c:/example/example/datfile.dat"))
}
