% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadingDatFiles.R
\name{ValidateMove}
\alias{ValidateMove}
\title{Validates that dats.split moved all files and headers are valid}
\usage{
ValidateMove(SourceFolder, DestinationFolder)
}
\arguments{
\item{SourceFolder}{The folder with .dat files needing splitting}

\item{DestinationFolder}{The folder where split .dat files should be written to}
}
\description{
LOWER LEVEL FUNCTION - Not meant to be called on its own. \cr \cr
Calls \code{\link{validate.dat.header}} on each file in the destination folder to assure all validation checks are passed.
It also confimrs that each Employer Id was moved.
}
\examples{
ValidateMove("c:/example/source/", "c:/example/destination/")
}
